Allow float in VerticalItemDecoration constructor dp parameter

+++ utils/src/main/java/com/baseconfig/pillar/utils/RecyclerViewUtils.java
@@ -17,7 +17,7 @@ public class RecyclerViewUtils {
             this.pixelMargin = pixelMargin;
         }
 
-        public VerticalItemDecoration(Context context, int dpMargin) {
+        public VerticalItemDecoration(Context context, float dpMargin) {
             this.pixelMargin = ContextUtils.getPixels(context, dpMargin);
         }
